...
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@12343 a1ca3aef-8c08-0410-bb20-df032aa958be

+++ cgi-bin/help-index.c
@@ -884,7 +884,7 @@ help_load_file(
 
 
   DEBUG_printf(("2help_load_file(hi=%p, filename=\"%s\", relative=\"%s\", "
-                "mtime=%ld)", hi, filename, relative, mtime));
+                "mtime=%ld)", hi, filename, relative, (long)mtime));
 
   if ((fp = cupsFileOpen(filename, "r")) == NULL)
     return (-1);
